fix(refresh): do not delete keyring password if network unavailable
If we run aws-creds without being connected to the internet the Okta API
call will return a network error. In the event of an error we naively
assume that our credentials are wrong and delete them from the system
keyring. There is a possibility that the error was caused due to
conenctivity issues and the credentials are actually OK. This change
detects if the current error is a net.OpError and if so returns an
explanatory messages and skips deleting the information from the system
keyring.